+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Overrode pbkdf2 package and upgraded related packages to fix critical audit issue
 - **CUMULUS-4039**
   - Updated to allow slashes in provider name and collection name 
+- **CUMULUS-4122**
+  - Change Granule ID search to be prefix by default. Added `Search by Infix` toggle option to Granule ID search.
 
 ## [v13.2.0] - 2025-05-14